Make sure Pcd service driver can work on 64-bit architecture.
git-svn-id: https://edk2.svn.sourceforge.net/svnroot/edk2/trunk@439 6f19259b-4bc3-4df7-8a09-765794883524

+++ edk2/EdkModulePkg/Universal/PCD/Dxe/Pcd.c
@@ -74,7 +74,7 @@ PcdDxeInit (
   //
   // Make sure the Pcd Protocol is not already installed in the system
   //
-  
+
   ASSERT_PROTOCOL_ALREADY_INSTALLED (NULL, &gPcdProtocolGuid);
   
   BuildPcdDxeDataBase ();
@@ -274,7 +274,7 @@ DxePcdGetSizeEx (
   IN UINTN                 ExTokenNumber
   )
 {
-  return DxePcdGetSize(GetExPcdTokenNumber (Guid, ExTokenNumber));
+  return DxePcdGetSize(GetExPcdTokenNumber (Guid, (UINT32) ExTokenNumber));
 }
